requirements.txt
yt-dlp>=2023.12.30
Pillow>=10.0.0
PyInstaller>=5.0.0